backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
pydantic-settings
typing-extensions
python-dotenv
langgraph
langchain
langchain-core
langchain-deepseek
langchain-openai
langchain-text-splitters
langchain-community
sentence-transformers
chromadb
faiss-cpu
pypdf
langsmith
numpy